add query execution time to server profiling system
