+++ Sources/IP/IP.swift
@@ -1,0 +1,4 @@
+@frozen public
+enum IP
+{
+}
